+++ UnityPackages/com.sizzle.abilitysystem/Runtime/Timeline/AbilityTimelineCommandType.cs
@@ -1,0 +1,13 @@
+namespace Sizzle.AbilitySystem.Timeline
+{
+    public enum AbilityTimelineCommandType
+    {
+        ActivateAbility,
+        CancelAbility,
+        CancelAllAbilities,
+        AddTag,
+        RemoveTag,
+        NotifyTag,
+        AddTimedTag,
+    }
+}